SLCORE-282 Change telemetry storage to use file locks
The previous implementation was keeping data in memory until upload, trying to merge with local storage latest changes in case of concurrent modifications.
This is not working well with counters, since merging is impossible.
Instead, any change is now written to disk synchronously, using a file lock to avoid concurrency issues (tested on Linux and Windows).

There is probbaly a performance cost to this, so we'll have to be careful to not increment counters too frequently.