Remove workaround for SonarTS
The TS sensor was manually excluded to handle the case where TS language was disabled but JS was enabled. This is not the case anymore as both are usually enabled at the same time, and the workaround does not work anymore since last SonarJS as both sensors have been merged.