SLCORE-562 Workaround to properly close SSE connection
It appera that cancelling the Future of the SSE request does not reliably close the underlying connection, leading
to receiving unexpected SSE events.

Waiting for a confirmation that there is an issue on HttpClient side, we found this workaround: throwing an excepting in the `data`
method will lead to the connection manager to release the connection.