Simplify waiting for the first branch matching
Shortly after startup, or lazyly on the first access, a branch matching will happen. Most other services are
interested by the result of this matching. We want to avoid the risk at startup for clients to return an empty branch
while in fact it is simply that the matching is still in progress.