SLCORE-737: Create OSGi bundle for Java Client
Using shading / relocating a OSGi bundle for Java Clients (e.g. Eclipse) was created. Due to conflicts in the Maven Shade and Bndtools plug-in a custom transformer for the META-INF/MANIFEST.MF file had to be created.

The OSGi bundle contains SLCORE libraries that are included but not shaded, some of them have packages that must be exported as well.