SLCORE-941: Don't fail on Windows Shortcuts (#1096)
Following the Windows specification we have to exclude Windows Shortcuts from the analysis as they are no links but actual, binary files.

The exclusion check if the file name hints to be a shortcut based on its extension. Only then we check the magic number and decide based on it.