SLCORE-1134 Fix flaky tests
Mockito throws random errors when methods are called on a spy at the same time it is being stubbed in another thread. To fix this, some tests were changed to stub before starting the backend. A trace was removed at the end of the backend creation to avoid using the spy