SLCORE-1141: Don't log an error when closed projects removed
Currently, when users close projects in the Eclipse IDE or when projects are removed on disk but the workspace reference is still in the workspace and the user then deletes this very project from the workspace, an error is logged.

As this is a valid situation in Eclipse, rather log a debug message and not an error as nothing is actually failing.